requirements.txt
apache-airflow==2.7.0
psycopg2-binary==2.9.9
pandas==2.0.3
requests==2.31.0